requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
duckduckgo-search>=5.0.0
wikipedia>=1.4.0
python-dotenv>=1.0.0